Refactor to be node compat
