+++ common.js
@@ -1,0 +1,5 @@
+const WARNING = "Warning";
+const ERROR = "Error";
+const RIGHT = "RIGHT";
+
+module.exports = { WARNING, ERROR, RIGHT };
